refactor: remove username page
