+++ client/src/views/OpenLogin.vue
@@ -1,28 +1,38 @@
 <template>
-  <b-container fluid class="h-100 w-100 d-flex justify-content-center align-items-center backdrop content-container">
+  <b-container fluid class="h-100 w-100 d-flex justify-content-center align-items-center backdrop">
     <div id="login-container" class="content-container">
-      <b-form>
-        <h1 class="text-center">Login or Sign Up</h1>
-        <hr>
-        <p>
-          Port of Mars is now in open beta, meaning anyone is welcome to sign up and
-          play as long you adhere to our <a href="" target="_blank">code of conduct</a>.
-        </p>
-        <hr>
-        <b-button :to="usernamePage" block variant="primary" size="lg" class="mb-3">
-          <b-icon icon="file-earmark-code" class="float-left"/>Sign in (Dev Mode)
-        </b-button>
-        <b-button block variant="white" size="lg" class="mb-3">
-          <img src="https://upload.wikimedia.org/wikipedia/commons/5/53/Google_%22G%22_Logo.svg" 
-            class="float-left" style="width:1.25em"/>Sign in with Google
-        </b-button>
-        <b-button block variant="facebook" size="lg" class="mb-3">
-          <b-icon icon="facebook" class="float-left"/>Sign in with Facebook
-        </b-button>
-        <b-button block variant="black" size="lg">
-          <b-icon icon="github" class="float-left"/>Sign in with Github
-        </b-button>
-      </b-form>
+      <h1 class="text-center">Login or Sign Up</h1>
+      <hr>
+      <p>
+        Port of Mars is now in open beta, meaning anyone is welcome to sign up and
+        play as long you adhere to our <a href="" target="_blank">code of conduct</a>.
+      </p>
+      <hr>
+      <div v-if="isDevMode" class="mb-3">
+        <b-form inline @submit="devLogin">
+          <b-form-input
+            id="input-username"
+            v-model="devLoginUsername"
+            placeholder="Dev Username"
+            required
+            class="w-50"
+          >
+          </b-form-input>
+          <b-button class="w-50" icon type="submit" variant="success" size="lg">
+            <b-icon class="mb-1" icon="box-arrow-right"></b-icon> Sign in
+          </b-button>
+        </b-form>
+      </div>
+      <b-button block variant="white" size="lg" class="mb-3">
+        <img src="https://upload.wikimedia.org/wikipedia/commons/5/53/Google_%22G%22_Logo.svg" 
+          class="float-left" style="width:1.25em"/>Sign in with Google
+      </b-button>
+      <b-button block variant="facebook" size="lg" class="mb-3">
+        <b-icon icon="facebook" class="float-left"/>Sign in with Facebook
+      </b-button>
+      <b-button block variant="black" size="lg">
+        <b-icon icon="github" class="float-left"/>Sign in with Github
+      </b-button>
     </div>
   </b-container>
 </template>
@@ -31,16 +41,34 @@
 import { Component, Vue } from "vue-property-decorator";
 import { url } from "@port-of-mars/client/util";
 import { isDevOrStaging } from "@port-of-mars/shared/settings";
-import { USERNAME_PAGE } from "@port-of-mars/shared/routes";
+import { DASHBOARD_PAGE, ONBOARDING_PAGE } from "@port-of-mars/shared/routes";
 
 @Component({})
 export default class OpenLogin extends Vue {
   isDevMode: boolean = false;
+  error: string = "";
+  devLoginUsername: string = "";
 
-  usernamePage = { name: USERNAME_PAGE };
+  dashboard = { name: DASHBOARD_PAGE };
+  usernamePage = { name: ONBOARDING_PAGE };
 
   async created() {
     this.isDevMode = isDevOrStaging();
+  }
+
+  async devLogin(e: Event) {
+    e.preventDefault();
+    const devLoginData: any = {
+      username: this.devLoginUsername,
+      password: "testing"
+    };
+    try {
+      await this.$ajax.devLogin(devLoginData);
+    } catch (e) {
+      if (e instanceof Error) {
+        this.error = e.message;
+      }
+    }
   }
 }
 </script>
